refactor(network): extract chrome entry to separate file
