+++ src/network/NetworkObservable.ts
@@ -2,31 +2,7 @@ import Protocol from 'devtools-protocol';
 import ProtocolMapping from 'devtools-protocol/types/protocol-mapping';
 import { ChromeRemoteInterface, Network } from 'chrome-remote-interface';
 import { Logger } from '../utils';
-
-export interface ChromeEntry {
-  isWebSocket: boolean;
-  frames?: {
-    type: 'request' | 'response';
-    time: number;
-    opcode: number;
-    mask: boolean;
-    data: string;
-  }[];
-  requestParams:
-    | Protocol.Network.RequestWillBeSentEvent
-    | Protocol.Network.WebSocketWillSendHandshakeRequestEvent;
-  responseParams?:
-    | { response: Protocol.Network.Response }
-    | Protocol.Network.ResponseReceivedEvent
-    | Protocol.Network.WebSocketHandshakeResponseReceivedEvent;
-  responseLength: number;
-  encodedResponseLength?: number;
-  responseFinishedS?: number;
-  responseFailedS?: number;
-  responseBody?: string;
-  responseBodyIsBase64?: boolean;
-  newPriority?: Protocol.Network.ResourcePriority;
-}
+import { ChromeEntry } from './ChromeEntry';
 
 export type ChromeRemoteInterfaceMethod = keyof ProtocolMapping.Events;
